bin: fix faulty assumption in locateCode and locateData
Prior to this commit, we relied on binary search to locate the first
candidate section in the list which could contain the code or data
searched for. However, as we searched for the end address, and not
the start address of the section, we did not satisfy the pre-condition
of binary search, as the list was sorted by start (and not end) section
address. Therefore our search property (end address) could result in
[false, true, false], while the binary search function sort.Search
requires a search property to result in a list with a clear boundary
between false and true results; e.g. [false, false, true].

Note to self: this issue was noticed in a PEF sample containing two
overlapping executable segments, one larger than the other. locateCode
returned the sorter one, and as such, the disassembler failed to locate
the underlying bytes of an address in an executable segment that should
have existed.